requirements.txt
aiocontextvars==0.2.2
aliyun-log-python-sdk==0.7.13
allure-pytest==2.11.1
allure-python-commons==2.11.1
altgraph==0.17.3
arrow==1.2.3
asgiref==3.3.4
asn1crypto==1.5.1
atomicwrites==1.4.1
attrs==22.1.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
blinker==1.4
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.5
contextvars==2.4
cryptography==2.4.2
dataclasses==0.8
dateparser==1.1.3
elastic-transport==8.4.0
elasticsearch==8.5.0
Faker==14.2.1
future==0.18.2
genson==1.2.2
gevent==22.10.2
greenlet==2.0.1
h11==0.13.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpath==0.82
jsonschema==3.2.0
kaitaistruct==0.8
ldap3==2.6.1
loguru==0.6.0
MarkupSafe==2.0.1
mitmdump==1.1.2
mitmproxy==5.0.0
msgpack==1.0.4
packaging==21.3
passlib==1.7.4
pefile==2022.5.30
pluggy==1.0.0
protobuf==3.10.0
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydivert==2.1.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyperclip==1.7.0
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
regex==2022.3.2
requests==2.27.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
six==1.16.0
sortedcontainers==2.1.0
SQLAlchemy==1.4.44
text-unidecode==1.3
tomli==1.2.3
tornado==6.1
typing_extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
urwid==2.0.1
Werkzeug==1.0.1
win32-setctime==1.1.0
wsproto==0.14.1
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
zstandard==0.12.0
